+++ artworks/constants.py
@@ -1,0 +1,7 @@
+"""Constants."""
+
+BATCH_SIZE = 10
+MONGODB_URL = "mongodb+srv://artist:music@bmat.jl51a.mongodb.net/artworks?retryWrites=true&w=majority"
+DATASET = 'artworks_db'
+COLLECTION = 'artworks'
+COLLECTION_TEST = 'artworks_test'