+++ CHANGELOG.md
@@ -6,5 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2021-03-03
+- Read from CSV and insert into MongoDB script
+- API with GET routes: to get right owners metadata and all artworks
+- Documentation
+
 ## [0.0.1] - 2021-03-02
 - Skeleton