+++ docs/changelog.rst
@@ -4,6 +4,12 @@ Changelog
 [Unreleased]
 ------------
 
+[1.0.0] - 2021-03-03
+--------------------
+* Read from CSV and insert into MongoDB script
+* API with GET routes: to get right owners metadata and all artworks
+* Documentation
+
 [0.0.1] - 2021-03-02
 --------------------
 * Skeleton